requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Werkzeug==3.1.3
python-dotenv==1.0.1
